+++ Development/KuditFrameworksTests/KuditFrameworksTests.swift
@@ -1,0 +1,16 @@
+//
+//  KuditFrameworksTests.swift
+//  KuditFrameworksTests
+//
+//  Created by Ben Ku on 6/19/24.
+//
+
+import Testing
+
+struct KuditFrameworksTests {
+
+    @Test func testExample() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+
+}